pyproject.toml
[tool.poetry]
name = "coleta-showtrials"
version = "0.2.0"
description = "Sistema de coleta, gestão e tradução de documentos históricos"
authors = ["Thiago Ribeiro <mackandalls@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]  # <-- ADICIONAR ESTA LINHA!

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
requests = ">=2.32.5"
beautifulsoup4 = ">=4.14.3"
lxml = ">=6.0.2"
rich = ">=13.7.0"
google-cloud-translate = ">=3.0.0"
python-dotenv = ">=1.0.0"
pytest = ">=9.0.0"  # <-- ADICIONAR PARA TESTES
fastapi = "^0.129.0"
uvicorn = "^0.40.0"
jinja2 = "^3.1.6"
aiofiles = "^25.1.0"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^7.0.0"
ruff = "^0.15.1"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]  # <-- ADICIONAR PARA TESTES!
[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = []
